+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.2.0
+
+* added `opts.prefix` option to `handleUpload` and renamed `uriSuffix` to `uri` in upload response
+* added jshint to tests
+
 ### 0.1.0
 
 * First version. Surely there will be more to come!